+++ dst/app/routes/accounts.js
@@ -21,6 +21,7 @@ const mongoose = require("mongoose");
 const authentication_1 = require("../middlewares/authentication");
 const permitScopes_1 = require("../middlewares/permitScopes");
 const validator_1 = require("../middlewares/validator");
+const USE_MONEY_TRANFER_AMOUNT_AS_NUMBER = process.env.USE_MONEY_TRANFER_AMOUNT_AS_NUMBER === '1';
 const accountsRouter = express_1.Router();
 const debug = createDebug('pecorino-api:router');
 const MAX_NUM_ACCOUNTS_CREATED = 100;
@@ -180,7 +181,14 @@ accountsRouter.get('/:accountType/:accountNumber/actions/moneyTransfer', permitS
         const searchConditions = Object.assign(Object.assign({}, req.query), { 
             // tslint:disable-next-line:no-magic-numbers
             limit: (req.query.limit !== undefined) ? Math.min(req.query.limit, 100) : 100, page: (req.query.page !== undefined) ? Math.max(req.query.page, 1) : 1, accountNumber: req.params.accountNumber });
-        const actions = yield actionRepo.searchTransferActions(searchConditions);
+        let actions = yield actionRepo.searchTransferActions(searchConditions);
+        // 互換性維持対応
+        if (USE_MONEY_TRANFER_AMOUNT_AS_NUMBER) {
+            actions = actions.map((a) => {
+                var _a;
+                return Object.assign(Object.assign({}, a), { amount: (typeof a.amount === 'number') ? a.amount : Number((_a = a.amount) === null || _a === void 0 ? void 0 : _a.value) });
+            });
+        }
         res.json(actions);
     }
     catch (error) {