+++ dst/app/routes/actions.js
@@ -19,6 +19,7 @@ const mongoose = require("mongoose");
 const authentication_1 = require("../middlewares/authentication");
 const permitScopes_1 = require("../middlewares/permitScopes");
 const validator_1 = require("../middlewares/validator");
+const USE_MONEY_TRANFER_AMOUNT_AS_NUMBER = process.env.USE_MONEY_TRANFER_AMOUNT_AS_NUMBER === '1';
 const actionsRouter = express_1.Router();
 actionsRouter.use(authentication_1.default);
 /**
@@ -63,7 +64,14 @@ actionsRouter.get('/moneyTransfer', permitScopes_1.default(['admin']), ...[
             // tslint:disable-next-line:no-magic-numbers
             limit: (req.query.limit !== undefined) ? Math.min(req.query.limit, 100) : 100, page: (req.query.page !== undefined) ? Math.max(req.query.page, 1) : 1 });
         const actionRepo = new pecorino.repository.Action(mongoose.connection);
-        const actions = yield actionRepo.searchTransferActions(searchConditions);
+        let actions = yield actionRepo.searchTransferActions(searchConditions);
+        // 互換性維持対応
+        if (USE_MONEY_TRANFER_AMOUNT_AS_NUMBER) {
+            actions = actions.map((a) => {
+                var _a;
+                return Object.assign(Object.assign({}, a), { amount: (typeof a.amount === 'number') ? a.amount : Number((_a = a.amount) === null || _a === void 0 ? void 0 : _a.value) });
+            });
+        }
         res.json(actions);
     }
     catch (error) {