+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Changed
 
+- MoneyTransferアクションのamountがMonetaryAmount型の場合に対応
+
 ### Deprecated
 
 ### Removed